requirements.txt

# FastAPI and ASGI server
fastapi==0.112.2
uvicorn[standard]==0.22.0
gunicorn==22.0.0
psycopg2==2.9.10
psycopg-binary==3.2.3
pydantic==2.8.2
pony==0.7.19

# JWT
PyJWT==2.9.0
passlib[bcrypt]==1.7.4

# Firebase Admin SDK
firebase-admin==6.1.0

# Database ORM and migrations
sqlalchemy==2.0.15
alembic==1.11.1
pymysql==1.0.3
# pymysql==1.0.3

# Environment variables
python-dotenv==1.0.0

# HTTP client for testing and external API calls
httpx==0.24.1

# Background tasks and job queuing (optional, include if needed)
# celery==5.3.0
redis== 5.0.8

# Logging (optional, as Python's built-in logging might be sufficient)
# loguru==0.7.0

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0

# Code formatting and linting (development dependencies)
black==23.3.0
flake8==6.0.0
isort==5.12.0

# Performance monitoring (optional)
# sentry-sdk==1.25.0

# AWS SDK (if you're using AWS services, optional)
# boto3==1.26.145


# Date and time handling (included in Python standard library)
# python-dateutil==2.8.2